refact. 대시보드 레이아웃 리팩토링
